Initial add of tool for keeping transmission up-to-date with server
The end goal is a client that keeps track of the correct torrent
list and has mutual TLS authentication, at which time this will
be retired.  As such, this piece is in python, such that I don't
waste a bunch of time on it.  python3 expected, REQUIREMENTS.txt
necessary to get a venv up-and-running included.